(mostly untested) Refactor control experiment names and update related functions
- Renamed control experiment types from "dumb_control" and "variation_control" to "same_data_control" and "same_distribution_control" across multiple files for consistency.
- Updated function names and references in `build_experiments.py`, `experiment_hf.py`, and `experiment.py` to reflect the new naming convention.
- Adjusted documentation and comments to align with the new experiment names.
- Modified plotting and result extraction scripts to accommodate the updated control experiment names.
- Added a new function to load pretrained models without fine-tuning for better flexibility in model usage.
- Enhanced activation extraction and probing scripts to support new experiment naming and ensure compatibility with existing workflows.

+++ dino/extract.py
@@ -75,7 +75,9 @@ def extract_activations_from_images(
             logits = feats["logits"].reshape(len(chunk), model.config.num_chars, model.config.num_classes)
             pred_chunks.append(logits.argmax(dim=-1).cpu().numpy())
 
-    activations = {name: np.concatenate(chunks) for name, chunks in accumulated.items()}
+    # float16 halves storage vs float32 with negligible effect on linear probe accuracy.
+    activations = {name: np.concatenate(chunks).astype(np.float16)
+                   for name, chunks in accumulated.items()}
     predictions = np.concatenate(pred_chunks)
     return activations, predictions
